chore: Sync Repository Labels (#28)
# Pull Request

## Description

This pull request introduces a new system for managing GitHub labels and
automates the synchronization of these labels using GitHub Actions. The
most important changes include the addition of a comprehensive list of
labels and the creation of a workflow to sync these labels.

Label management:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR1-R96):
Added a detailed list of default GitHub labels, custom labels, and
labels for different components and sizes.

Automation:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6R1-R30):
Created a GitHub Actions workflow to automatically sync labels from the
`labels.yml` file whenever changes are pushed to the `main` branch.

Fixes #17