ci: Run Validator before Deploy (#35)
# Pull Request

## Description


This pull request includes significant changes to the
`.github/workflows/deploy.yml` file to introduce a new job for running a
repository standards validator and modifying the existing build and
deploy jobs to accommodate this new step. The most important changes
include the addition of the `run-validator` job, updates to the
`build-site` job to depend on the validator, and changes to the `deploy`
job dependencies.

Changes to workflow jobs:

* Added a new `run-validator` job to run a repository standards
validator. This job includes steps to checkout the repository, set up
Python, install dependencies, run the validator, and upload the results.

Modifications to existing jobs:

* Updated the `build-site` job to depend on the `run-validator` job and
added steps to download the validator results.
* Changed the `deploy` job to depend on the `build-site` job instead of
the `build` job.

Fixes #36