chore: Configure ESLint to play nice with Prettier (#49)
# Pull Request

## Description

This pull request includes updates to the ESLint configuration to
integrate Prettier for code formatting. The most important changes
include importing the `eslint-config-prettier` package, adding it to the
ESLint configuration, and updating the `package.json` file to include
the new dependency.

### ESLint Configuration Updates:

*
[`eslint.config.js`](diffhunk://#diff-a32a0887ed9d1d707bbb3b845b7df7fd40e673c47e7b60a3ebd896b68d3b8839R5):
Added import for `eslint-config-prettier` and included it in the ESLint
configuration array to integrate Prettier for code formatting.
[[1]](diffhunk://#diff-a32a0887ed9d1d707bbb3b845b7df7fd40e673c47e7b60a3ebd896b68d3b8839R5)
[[2]](diffhunk://#diff-a32a0887ed9d1d707bbb3b845b7df7fd40e673c47e7b60a3ebd896b68d3b8839R19)

### Dependency Updates:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R32):
Added `eslint-config-prettier` version `9.1.0` to the list of
dependencies.