ci: Add CodeQL Analysis (#68)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow for
performing CodeQL analysis. The most important changes include the
addition of the `run-codeql-analysis` job, which is configured to
analyze Python and JavaScript code for security and quality issues.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R122-R144):
Added `run-codeql-analysis` job to perform CodeQL analysis on Python and
JavaScript code. This job includes steps for checking out the
repository, initializing CodeQL, and performing the analysis.

Fixes #20