chore: Add Code Of Conduct (#72)
# Pull Request

## Description

This pull request introduces a new Code of Conduct for the project. The
document outlines the community's commitment to creating a welcoming and
inclusive environment, sets standards for acceptable behavior, and
details the enforcement procedures for violations.

Key additions include:

* **Our Pledge**: A commitment to providing a harassment-free experience
for everyone, regardless of personal characteristics.
* **Our Standards**: Examples of positive behavior and unacceptable
behavior.
* **Enforcement Responsibilities**: Clarification of the
responsibilities of community leaders in enforcing the Code of Conduct.
* **Scope**: The application of the Code of Conduct within all community
spaces and when representing the community in public spaces.
* **Enforcement Guidelines**: Detailed consequences for violations,
ranging from a correction to a permanent ban.

Fixes #6