chore: Add Conventional Commits (#73)
# Pull Request

## Description

This pull request introduces a detailed documentation file for the
Conventional Commits specification. The document outlines the structure,
types, and benefits of using Conventional Commits, along with examples
and a FAQ section to help developers understand and implement the
specification effectively.

Key additions to the documentation:

*
[`docs/CONVENTIONAL_COMMITS.md`](diffhunk://#diff-ec1fec2852f28b94693945c6e635f6b64911318d47257951cdfd0295b11d004cR1-R205):
Added a comprehensive guide on Conventional Commits, including the
summary, commit message structure, types of commits, examples, and
detailed specification. The document also covers the benefits of using
Conventional Commits and provides a FAQ section to address common
questions and scenarios.

Fixes #69