chore: Fix Data Table Pagination (#85)
# Pull Request

## Description

This pull request refactors the code to improve modularity and
maintainability by moving the `RepositoryTable` component into its own
file and updating the relevant imports. The most important changes
include the creation of the new `RepositoryTable` component, and the
corresponding updates to the `index.astro` file to use this new
component.

Refactoring and modularization:

*
[`dashboard/src/components/RepositoryTable.tsx`](diffhunk://#diff-deff929dd0843abdf2ec302d74cda25106f47341d30bc84b69af73b43ec0f97fR1-R10):
Created a new `RepositoryTable` component that imports `DataTable`,
`columns`, and `repositories` and returns a `DataTable` with the
appropriate columns and data.
*
[`dashboard/src/pages/index.astro`](diffhunk://#diff-4698539578136e340030d4a0beaf14c88eb652d6090c3a9f906c414a3557fc8bL4-R4):
Updated to import the newly created `RepositoryTable` component instead
of directly importing `DataTable`, `columns`, and `repositories`.
[[1]](diffhunk://#diff-4698539578136e340030d4a0beaf14c88eb652d6090c3a9f906c414a3557fc8bL4-R4)
[[2]](diffhunk://#diff-4698539578136e340030d4a0beaf14c88eb652d6090c3a9f906c414a3557fc8bL17-R15)

Fixes #84