build: Fix Homepage Layout (#103)
# Pull Request

## Description

This pull request includes updates to the ESLint configuration and
changes to the HTML class attribute syntax in the `dashboard` project.
The most important changes are:

### ESLint Configuration Updates:
*
[`dashboard/eslint.config.js`](diffhunk://#diff-ba3688074c7e0e5c392c81b0a0c78e44a1aafc568b73a3d2d13170a87130fc06R19):
Disabled the `react/no-unknown-property` rule to prevent linting errors
for custom properties in React components.

### HTML Class Attribute Syntax Changes:
*
[`dashboard/src/pages/index.astro`](diffhunk://#diff-4698539578136e340030d4a0beaf14c88eb652d6090c3a9f906c414a3557fc8bL8-R14):
Changed the `className` attribute to `class` in multiple HTML elements
to ensure compatibility with Astro framework.

Fixes #102