build: Link Repository Name to GitHub Page (#105)
# Pull Request

## Description

This pull request includes changes to the
`dashboard/src/components/columns.tsx` file to enhance the display of
repository information in the dashboard. The most important changes
include adding a new property to the `Repository` type and updating the
column definition to create clickable links for repository names.

Enhancements to repository display:

*
[`dashboard/src/components/columns.tsx`](diffhunk://#diff-911097980600deaf4c658f70d24ffc5cde6d9e393aec77f360241c34507341a3R9):
Added a new property `repository_link` to the `Repository` type to store
the URL of the repository.
*
[`dashboard/src/components/columns.tsx`](diffhunk://#diff-911097980600deaf4c658f70d24ffc5cde6d9e393aec77f360241c34507341a3L18-R30):
Updated the `name` column definition to render the repository name as a
clickable link that opens in a new tab.

Fixes #104