test: Add robots.txt Test (#109)
# Pull Request

## Description

This pull request introduces a new test for checking the accessibility
of the `robots.txt` file and updates the constants accordingly. The most
important changes include the addition of a new constant for the
`robots.txt` URL and the implementation of a new test function.

### New Test Implementation:

*
[`tests/ui/test_dashboard.py`](diffhunk://#diff-0f5c761cdc6013814735e5a86f696efa96ebec90d540a4517d2deb396fdefbf7R35-R44):
Added a new test function `test_robots` to check the accessibility and
content of the `robots.txt` file.

### Constants Update:

*
[`tests/ui/utils/constants.py`](diffhunk://#diff-ffd2fa12c0a17c2250e4cebd95863d383a45bb9190556d24a4bff98568d1562bR5):
Added a new constant `ROBOTS_URL` to store the URL of the `robots.txt`
file.

### Import Update:

*
[`tests/ui/test_dashboard.py`](diffhunk://#diff-0f5c761cdc6013814735e5a86f696efa96ebec90d540a4517d2deb396fdefbf7L4-R4):
Updated the import statement to include the new `ROBOTS_URL` constant.

Fixes #97