test: Add Test for Sorting Table (#115)
# Pull Request

## Description

This pull request introduces new functionality and improvements to the
test suite. The most important changes include adding a new command to
run UI tests in headed mode and introducing a test for table sorting in
the dashboard.

### New functionality:

*
[`tests/tests.just`](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918L13-R20):
Added a new command `ui-tests-headed` to run UI tests in headed mode.

### Test improvements:

*
[`tests/ui/test_dashboard.py`](diffhunk://#diff-0f5c761cdc6013814735e5a86f696efa96ebec90d540a4517d2deb396fdefbf7R45-R66):
Added a new test `test_table_sorting` to verify that the repository
table can be sorted in ascending and descending order.

Fixes #114