build: Sort All Table Columns (#124)
# Pull Request

## Description

This pull request includes changes to the
`dashboard/src/components/columns.tsx` file to enhance the header
functionality of several columns by adding sorting capabilities.

Enhancements to column headers:

* Added a `Button` component to the `secret_scanning_push_protection`
column header to enable sorting when the header is clicked.
* Added a `Button` component to the `secret_scanning` column header to
enable sorting when the header is clicked.
* Added a `Button` component to the `dependabot_security_updates` column
header to enable sorting when the header is clicked.

Fixes #122