build: Update Button Colours (#161)
# Pull Request

## Description

This pull request includes several changes to the
`dashboard/src/components` files, focusing on improving the visual
representation of data in the `Badge` components and updating icon
usage. The key changes include modifying the `Badge` component's class
names based on conditions and updating the icon sizes and imports.

### Visual Improvements:

*
[`dashboard/src/components/RepositoryDetailColumns.tsx`](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eL55-R54):
Updated the `Badge` component's class to change the background color
based on the number of open pull requests and issues.
[[1]](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eL55-R54)
[[2]](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eL70-R74)

### Icon Updates:

*
[`dashboard/src/components/RepositoryDetailColumns.tsx`](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eL6-R6):
Removed unused icons and updated the import statement to include only
the necessary icons.
*
[`dashboard/src/components/RepositoryKeyFilesColumns.tsx`](diffhunk://#diff-9022d5265d034e4497814de03ff0de7f79d6fed32ca95cc17422a9c255a6af4cL6-R11):
Updated the size of the `CheckIcon` and `CrossIcon` and replaced
`CrossIcon` with `X` for better clarity.

### Security Column Updates:

*
[`dashboard/src/components/RepositorySecurityColumns.tsx`](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L52-R55):
Modified the `Badge` component's class names to use specific background
colors instead of variant names for various security-related columns.
[[1]](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L52-R55)
[[2]](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L78-R82)
[[3]](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L98-R104)
[[4]](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L122-R127)
[[5]](diffhunk://#diff-1bbedce9e1f24ae279a33dd5a6fe34679450ab626c98133529544bbf77d6e993L148-R154)