Add Issue Templates (#169)
# Pull Request

## Description

This pull request includes changes to the issue templates to standardize
and improve the process of reporting bugs and requesting features. The
most important changes include the addition of new templates for bug
reports and feature requests, and the disabling of blank issues.

Standardization of issue templates:

*
[`.github/ISSUE_TEMPLATE/bug_report.md`](diffhunk://#diff-185833cb26d7ac66a4d39042fd576a820c2c2c6d05ad18973bb9c7dce77267c5R1-R33):
Added a new template for bug reports to ensure consistent and detailed
reports from users.
*
[`.github/ISSUE_TEMPLATE/feature_request.md`](diffhunk://#diff-148870715557a13127284f8aeaa28002ed6b034268af13c5d030ab59fd078220R1-R19):
Added a new template for feature requests to gather clear and concise
ideas from users.

Configuration updates:

*
[`.github/ISSUE_TEMPLATE/config.yml`](diffhunk://#diff-1c0d972ee49103af56fd608a77a28e1eb12f6908f263f0f183d46868fdcd8ea2R1):
Disabled the ability to create blank issues to encourage the use of the
provided templates.

Fixes #9