build: Link Issues Count to GitHub Tab (#176)
# Pull Request

## Description

This pull request includes changes to the `RepositoryDetailColumns`
component in the `dashboard/src/components/RepositoryDetailColumns.tsx`
file. The main update is the addition of a hyperlink to the repository
issues page within the `cell` rendering logic.

Enhancements to the `RepositoryDetailColumns` component:

* Added an anchor tag around the `Badge` component to link to the
repository's issues page. This makes the number of open issues
clickable, directing users to the issues page in a new tab.
(`dashboard/src/components/RepositoryDetailColumns.tsx`)
[[1]](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eR75-R80)
[[2]](diffhunk://#diff-e1cf8cee5106901b7bc15f9193d142819cca9797beae05943113409a644d251eR89)

Fixes #174