build: Add Logo (#179)
# Pull Request

## Description

This pull request includes updates to the `README.md` file and changes
to the favicon in the `Layout.astro` file.

Documentation update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R1-R8):
Added a new section for the `repo-overseer` project, including a brief
description and a table of contents.

User interface update:

*
[`dashboard/src/layouts/Layout.astro`](diffhunk://#diff-5f757a7d2e52d2bc5226cc6d64fbcf9b17fa90e57b96c550f5c2ec426c3eaa76L8-R9):
Changed the favicon type from `image/svg+xml` to `image/x-icon` and
updated the favicon file reference from `favicon.svg` to `favicon.ico`.

Fixes #165