test: Add Tab Tests (#183)
# Pull Request

## Description

This pull request includes various updates and additions to the test
suite, primarily focusing on the UI tests for different tabs on the
dashboard page. The changes also include the removal of a redundant test
and the renaming of a test file for better clarity.

### Additions to UI tests:

*
[`tests/ui/test_dashboard_details.py`](diffhunk://#diff-449c8894c23cc82e570b5b695ea79e6bcf07a98992d29de37d92d9a525730d08R1-R31):
Added a new test to verify the sorting functionality of the repository
table on the details tab.
*
[`tests/ui/test_dashboard_key_files.py`](diffhunk://#diff-dda22b85ff93083cd16752c35b23272b58376a932a83adaf32ef5ed1d57ffcaeR1-R31):
Added a new test to verify the sorting functionality of the repository
table on the key files tab.
*
[`tests/ui/test_dashboard_security.py`](diffhunk://#diff-228e9019ceded0e4dacaccce29252eaf2079ef2ea1e4f4c5634aee2afa98057eR1-R31):
Added a new test to verify the sorting functionality of the repository
table on the security tab.

### Renaming and cleanup:

*
[`tests/ui/test_dashboard_site_wide.py`](diffhunk://#diff-ed1063b326c145e2375444e5b49a7b7da16340827d8025eeed70e34b53ff0100R1-R2):
Renamed the file from `test_dashboard.py` and removed a redundant
sorting test.
[[1]](diffhunk://#diff-ed1063b326c145e2375444e5b49a7b7da16340827d8025eeed70e34b53ff0100R1-R2)
[[2]](diffhunk://#diff-ed1063b326c145e2375444e5b49a7b7da16340827d8025eeed70e34b53ff0100L45-L66)

### Miscellaneous:

*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL28):
Removed the `D100` ignore rule from the list of ignored checks.
*
[`tests/ui/utils/constants.py`](diffhunk://#diff-ffd2fa12c0a17c2250e4cebd95863d383a45bb9190556d24a4bff98568d1562bR1-R2):
Added a docstring to the constants module.

Fixes #182