chore: Rename Action Steps (#195)
# Pull Request

## Description

This pull request includes changes to the `.github/workflows/deploy.yml`
file to add names to the jobs in the GitHub Actions workflow. These
changes improve the readability and clarity of the workflow by providing
descriptive names for each job.

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R15):
Added the name `Run Validator` to the `run-validator` job.
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R47):
Added the name `Build Site` to the `build-site` job.
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R70):
Added the name `Deploy to GitHub Pages` to the `deploy` job.