ci: Update Dependency Review (#215)
# Pull Request

## Description

This pull request includes updates to the dependency review job in the
GitHub Actions workflow configuration file. The changes enhance the
permissions and improve the feedback mechanism for dependency reviews.

Enhancements to dependency review job:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR50-R51):
Added `write` permissions for `pull-requests` to the `dependency-review`
job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR60-R61):
Configured the `dependency-review-action` to comment a summary in the
pull request on failure.