chore: Fix Some UI Tests (#217)
# Pull Request

## Description

This pull request includes changes to the UI tests to add a new test
command, import a necessary module, and modify the table sorting test.
The most important changes are summarized below:

New test command:

*
[`tests/tests.just`](diffhunk://#diff-83e2fb4528694cfa62c213eefec53e6aeee0353030bf91ad4f830e69854b7918R25-R28):
Added a new command `ui-tests-wip` to run UI Playwright tests in headed
mode with the "wip" keyword.

Test improvements:

*
[`tests/ui/test_common_table.py`](diffhunk://#diff-34f38616313a0f75a060ce95eca333cdba3d77967f1768d29b8c0c3b61ea322eR3-R4):
Imported the `sleep` function from the `time` module to introduce a
delay in the table sorting test.
*
[`tests/ui/test_common_table.py`](diffhunk://#diff-34f38616313a0f75a060ce95eca333cdba3d77967f1768d29b8c0c3b61ea322eR16-R31):
Added a `sleep(1)` call before waiting for the table to be loaded
initially in the `test_table_sorting__repository_column` function.
*
[`tests/ui/test_common_table.py`](diffhunk://#diff-34f38616313a0f75a060ce95eca333cdba3d77967f1768d29b8c0c3b61ea322eR16-R31):
Modified the `test_table_sorting__repository_column` function to use
`page.get_by_role("button", name="Repository").click()` directly instead
of assigning it to a variable.
*
[`tests/ui/test_common_table.py`](diffhunk://#diff-34f38616313a0f75a060ce95eca333cdba3d77967f1768d29b8c0c3b61ea322eR16-R31):
Updated the expected text in the first cell after sorting to
"actions-status" with a timeout of 5000 milliseconds.