chore: Update SuperLinter EditorConfig Checker (#250)
# Pull Request

## Description

This pull request includes changes to the configuration and workflow
files to correct a typo and update the linter settings.

Configuration and workflow updates:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L5-R5):
Corrected a typo in the `SpacesAfterTabs` setting and renamed the file
from `.ecrc` to `.editorconfig-checker.json`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R35):
Added `EDITORCONFIG_FILE_NAME` to the environment variables to specify
the new name of the editor configuration file.