ci: Add Pinact Verify Job (#274)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
verify the Pinact configuration. The most important change is the
introduction of the `pinact-verify` job in the
`.github/workflows/code-checks.yml` file.

Additions to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R220-R233):
Introduced a new job named `pinact-verify` that runs on `ubuntu-latest`.
This job includes steps to check out the repository, install the latest
version of `uv` using the `astral-sh/setup-uv` action, and ensure proper
configuration verification.