ci: Remove Pinact Verify (#278)
# Pull Request

## Description

This pull request includes two main changes: the removal of the
`pinact-verify` job from the GitHub Actions workflow and the update of
the `uv` tool's required version in the Python project's
`pyproject.toml` file.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L236-L249):
Removed the `pinact-verify` job, which included steps for checking out
the repository and installing the latest version of `uv`.

### Dependency updates:
*
[`tests/pyproject.toml`](diffhunk://#diff-03049af3cb225ee5c5f22bc10731df89191cc762ad9ca9aa4cb7d3a87786ac8dL11-R11):
Updated the `uv` tool's `required-version` from `0.6.16` to `0.6.17`.