chore: Fix Zizmor SARIF Command (#309)
# Pull Request

## Description

This pull request updates the workflow and build process to replace the
setup and execution of Python dependencies with new tools (`uv` and
`Just`) and modifies the commands used for running and uploading SARIF
files. The most important changes are grouped below by theme.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L227-R236):
Replaced the setup of Python dependencies with the installation of the
latest version of `uv` and the setup of `Just` using their respective
GitHub actions. Updated the command for running `zizmor` to use `just
zizmor-check-sarif` and upgraded the `codeql-action/upload-sarif` action
to version `v3.28.17`.

### Build process updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL53-R57):
Added a new `zizmor-check-sarif` recipe that runs `zizmor` with SARIF
output using `uvx`. Updated the existing `zizmor-check` recipe to use
`uvx` instead of the previous command.