chore: Fix Zizmor Warnings (#362)
# Pull Request

## Description

This pull request includes a small update to the
`.github/workflows/deploy.yml` file. The change modifies how the browser
environment variable is passed to the UI tests in the workflow.

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L93-R95):
Updated the `Run UI Tests` step to use an explicit `BROWSER` environment
variable instead of directly referencing `matrix.browser` in the
command.