ci: Swap linkinator to linkcheck (#385)
# Pull Request

## Description

This pull request updates the link-checking tool in the GitHub Actions
workflow to a newer and more feature-rich version, along with changes to
its configuration.

### Workflow updates:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L108-R110):
Replaced the `linkinator-action` tool with `linkcheck` for running link
tests, upgrading from version `1.11.0` to `3.0.0`. Updated the
configuration to use the `arguments` parameter instead of `paths`, and
removed additional parameters such as `recurse`, `timeout`, and
`markdown`.