chore: Fix UI Tests (#489)
# Pull Request

## Description

This pull request makes a minor update to a UI test to reflect a change
in the expected number of table rows after pagination. The assertion now
expects 3 rows instead of 2.

- Updated the assertion in `test_details_table_pagination` in
`test_common_table.py` to expect 3 table rows after pagination, likely
due to a change in the underlying data or pagination logic.