ci: Update Pull Request Tasks Permissions (#491)
# Pull Request

## Description

This pull request updates the permissions configuration in the
`.github/workflows/pull-request-tasks.yml` workflow. The main change is
to scope permissions more narrowly for the `check-pull-request-title`
job.

Workflow permissions updates:

* Replaced the global `contents: read` permission with an empty
permissions object at the workflow level, and set `pull-requests: read`
permission specifically for the `check-pull-request-title` job to follow
the principle of least privilege.