chore: Fix Trivy Warnings (#554)
# Pull Request

## Description

This pull request introduces a change to the dependency management
configuration in the `tests/pyproject.toml` file. The most important
update is the addition of an override for the `urllib3` package version,
which ensures that version 2.6.0 is used during testing.

Dependency management:

* Added `override-dependencies` for `urllib3==2.6.0` in the `[tool.uv]`
section of `tests/pyproject.toml` to enforce the use of this specific
version during test runs.